requirements.txt
asgiref==3.5.2
boto3==1.26.6
botocore==1.29.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
dynaconf==3.1.11
gunicorn==20.1.0
hvac==1.0.2
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pyhcl==0.4.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
shortuuid==1.0.10
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
whitenoise==6.2.0
zipp==3.10.0
